backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
chromadb==0.5.15
openai==1.54.3
pypdf2==3.0.1
python-dotenv==1.0.1
sqlalchemy==2.0.36
